Patch 12/2021 (#920)
* Remove ability to calculate bonds from species viewer, as this will break EVERYTHING.

* Update info frame point size in viewers so they're legible.

* Mark overrides.

* Remove unused generator group from SpeciesTab.

* Use label for applied size factor value.

* Fix info text.

* Fix transitions between module control widgets.

* Adjust main window margins.

* Fix limit calculation for periodic species.

* Don't add species if population is <= 0.

* Don't try to set parameters vector if it wasn't provided.

* Don't error and stop on missing file. Reposition module setUp().

* Tidy recent files functions.

* Don't enforce integer populations in Isotopologue.

* Remove some redundant declarations.

* Remove Q_DECLARE_* uses.

* Remove unused includes.

* Disable simulation menu when using Run For.

* Fix crash from bad keyword use in EPSRModule.

* Fix graphing in EPSRModule.

* Remove merge conflict markers in comment.

* Check for presence of atom in selection before trying to erase it.

* Return Isotopologues for any root item column.

* Update control widget when editing module name.

* Disable controls / features in LayerTab while running.

+++ src/classes/neutronweights.h
@@ -27,7 +27,7 @@ class NeutronWeights
     // Clear contents
     void clear();
     // Add Species Isotopologue to the relevant mixture
-    void addIsotopologue(const Species *sp, int speciesPopulation, const Isotopologue *iso,
+    void addIsotopologue(const Species *sp, double speciesPopulation, const Isotopologue *iso,
                          double isotopologueRelativePopulation);
     // Return whether an Isotopologues definition exists for the provided Species
     bool containsIsotopologues(const Species *sp) const;
